Added verification program to tool suite.
Largely rewrote the StreamBuffer class in an attempt to make it slightly
more readable. Moved the Producer and Consumer classes outside of the
ProducerConsumerModel to better decouple the classes. Added a gnuradar
verification problem to validate the user's desired sampling window size
and what actually results from hardware with the system's rx trigger.