Test server code is now functioning.
The initial version of the client/server code is now functioning as a
simple test. XML parsing has also been verified. The next step in the
saga is to add necessary commands and incorporate the client server in a
new program called gradar-run-daemon. This daemon should listen for
control/status requests on the proper port.