Updates to the installer.
The installer is still far from what I want, but scons should be
capable, it's just a matter of spending more time with it. The install now
includes java wrappers for gradar-configure and gradar-verify. The
command-line versions of these programs are now differentiated by a -cli
suffix. Binaries are now dumped in /usr/local/bin by default.

TODO:

Move the java xom dependency out of the lib subdirectory and drop it
next to the application's jar file.