Preliminary server-based run program implementation.
The foundation for a server-based run program has been designed and the
core has now been implemented. The run program will now accept commands
using a TCP/IP server that listens on a specific socket. Commands must
follow a predefined format provided in XML. The start command has been
implemented and tested as a simple proof-of-concept. In the days that
follow, all control and status functions will be implemented this way.
The end game is to provide a java-based GUI that will talk to the
server, either locally or on another machine connected by a network, and
send control commands, display system status, and provide basic run-time
diagnostics to ensure the radar is running smoothly.