Continued modifications and code required for networked data collection.
This commit includes 20 hours worth of modifications to re-organize the
code for networked data collection. The Producer/Consumer model had to
be totally reworked to work with the networked setup. The primary
difference was the old system would start the consumer and producer
threads, but one the consumer side was not ran as a thread, but a while
loop that would sleep when no data was available. This method was
perfectly fine ( it doesn't eat cpu cycles ) for the command-line
version, but, in the networked version, both consumer and producer must
operate completely indepedently, and the threads simply fall through and
the server blocks, waiting for a control/status commands from the
client. There are still quite a few issues to work out with regards to
stopping/starting the system and dealing with shared pointers, but a
simple test has been completed and noise samples were collected and
proplerly aligned.

This commit was tested ONLY in networked mode using libusrp-v3.3.

TODO:

1. Work out details to properly start/restart system without exiting and
   restarting the server. This will require some thought w.r.t. memory
   management and current use of shared pointers, which are sprinkled
   throughout the system ( and have side-effects when used blindy ).

2. Continue testing and coding to handle exceptions.