Tested version of gradar-run and gradar-run-server.
Basic testing is now complete for the two primary programs involved in a
networked setup. So far I've only tested both programs locally, but,
ideally, you should be able to start the server from the machine
attached to the hardware, and then run gradar-run on another computer
pointed to the server's location.

TODO:

1. Add more features to gradar-run.
2. Provide more test cases for completeness.
3. Provide information concerning port settings ( i.e. /etc/services ).
4. Provide a configuration file to configure the server's ip and port.