First attempt at a stand-alone stripped-down version of the USRP source.
The goal is to eventually remove complete dependence on the super-large
USRP library and associated modules. For this project, only the basic RX
board is support with the bare c++ interface ( i.e. usrp_standard ).
This will remove major pains due to reliance on autoconf and cousins
along with the sheer fact that gnuradio builds can take 15-20 minutes,
depending on the modules being built.