Fixed a bug in GnuRadarSettings initialization and added phase programming to the configuration.
1. In some instances, the Tune() command was called before setting the
actual number of channels used in the setup. This probably should be
handled via CTOR, but, for now, I'll simply fix the order of the two
instances that use the class.

2. Phase tuning was not included in previous builds. It should work now.

NOTES:

4-channel setups appear to have issues with higher bandwidth settings,
which point to FPGA timing issues. For now, I've been able to
successfully use 4-channel mode at a decimation of 32.