Fixed false triggering bug.
When attempting to use a delayed trigger, the system would falsely trigger and store empty data tables. The call to usrp.read() returns the number of bytes read, or an error given by -1. The system was not using this return value and it appears that some sort of timeout returns an error. This bug has been fixed in this version.